DSP-11 create api for allLanguages in xml
