+++ api.js
@@ -2,3 +2,4 @@ import { API } from "./api/api.js";
 
 export const lastUpdated = API.getLastUpdated();
 export const availableServices = API.getAllServiceNames();
+export const availableLanguages = API.getAllLanguages();
